+++ history.md
@@ -3,6 +3,8 @@
 
 ## v0.4.8 2024-01-14 ##
 * Replace urease with seqkit
+* Refine config of reference
+* Refine prepare_uniprot_taxonomy function
 
 ## v0.4.7 2022-12-05 ##
 * Added more parallel option for preprocessing workflow